+++ src/shared/request.rs
@@ -1,15 +1,15 @@
 use serde::Deserialize;
-use serde_with::DisplayFromStr;
 use serde_with::serde_as;
+use serde_with::DisplayFromStr;
 
 #[serde_as]
 #[derive(Debug, Deserialize)]
 pub struct ListQuery {
     #[serde_as(as = "Option<DisplayFromStr>")]
-    pub start: Option<i64>,
+    pub start: Option<i32>,
 
     #[serde_as(as = "Option<DisplayFromStr>")]
-    pub limit: Option<i64>,
+    pub limit: Option<i32>,
 
     pub keyword: Option<String>,
 
@@ -19,8 +19,8 @@ pub struct ListQuery {
 }
 
 pub trait ListQueryImpl {
-    fn start(&self) -> Option<i64>;
-    fn limit(&self) -> Option<i64>;
+    fn start(&self) -> Option<i32>;
+    fn limit(&self) -> Option<i32>;
     fn keyword(&self) -> Option<String>;
     fn sort_by(&self) -> Option<String>;
     fn order(&self) -> Option<String>;
